move gcs store to modules directory
