+++ src/store/modules/Election2020/index.js
@@ -2,8 +2,7 @@ import Vue from 'vue'
 import { get } from 'lodash'
 import { getSheet } from 'src/api'
 import { timerModule as timer } from './modules/timer'
-
-import gcs from './gcs.js'
+import gcs from './modules/gcs'
 
 export default {
   namespaced: true,
